+++ src/Data/Serialize/Describe/Combinators/LE.hs
@@ -2,11 +2,10 @@
 --
 -- All combinators take a function that takes the structure being described ('a') and produces the specified data type from it.
 -- Most of the time, this will be one of the structure's fields, which are all functions from the structure to the field type.
-
-
 module Data.Serialize.Describe.Combinators.LE(
-  w8, w16, w32, w64,
-  i8, i16, i32, i64,
+  LE(..),
+  w16, w32, w64,
+  i16, i32, i64,
   f32, f64
 ) where
 
@@ -16,33 +15,55 @@ import Data.Serialize.IEEE754
 import Data.Serialize.Get
 import Data.Serialize.Put
 import Data.Serialize.Describe.Descriptor
+import Data.Serialize.Describe.Class
 
-w8 :: (a -> Word8) -> Descriptor a Word8
-w8 f = Descriptor (getWord8, \s' -> putWord8 (f s') >> pure (f s'))
+w16 :: Integral i => (s -> i) -> Descriptor s i
+w16 f = Descriptor (fromIntegral <$> getWord16le, \s' -> putWord16le (fromIntegral $ f s') >> pure (f s'))
 
-w16 :: (a -> Word16) -> Descriptor a Word16
-w16 f = Descriptor (getWord16le, \s' -> putWord16le (f s') >> pure (f s'))
+w32 :: Integral i => (s -> i) -> Descriptor s i
+w32 f = Descriptor (fromIntegral <$> getWord32le, \s' -> putWord32le (fromIntegral $ f s') >> pure (f s'))
 
-w32 :: (a -> Word32) -> Descriptor a Word32
-w32 f = Descriptor (getWord32le, \s' -> putWord32le (f s') >> pure (f s'))
+w64 :: Integral i => (s -> i) -> Descriptor s i
+w64 f = Descriptor (fromIntegral <$> getWord64le, \s' -> putWord64le (fromIntegral $ f s') >> pure (f s'))
 
-w64 :: (a -> Word64) -> Descriptor a Word64
-w64 f = Descriptor (getWord64le, \s' -> putWord64le (f s') >> pure (f s'))
+i16 :: Integral i => (s -> i) -> Descriptor s i
+i16 f = Descriptor (fromIntegral <$> getInt16le, \s' -> putInt16le (fromIntegral $ f s') >> pure (f s'))
 
-i8 :: (a -> Int8) -> Descriptor a Int8
-i8 f = Descriptor (getInt8, \s' -> putInt8 (f s') >> pure (f s'))
+i32 :: Integral i => (s -> i) -> Descriptor s i
+i32 f = Descriptor (fromIntegral <$> getInt32le, \s' -> putInt32le (fromIntegral $ f s') >> pure (f s'))
 
-i16 :: (a -> Int16) -> Descriptor a Int16
-i16 f = Descriptor (getInt16le, \s' -> putInt16le (f s') >> pure (f s'))
+i64 :: Integral i => (s -> i) -> Descriptor s i
+i64 f = Descriptor (fromIntegral <$> getInt64le, \s' -> putInt64le (fromIntegral $ f s') >> pure (f s'))
 
-i32 :: (a -> Int32) -> Descriptor a Int32
-i32 f = Descriptor (getInt32le, \s' -> putInt32le (f s') >> pure (f s'))
+f32 :: (Real f, Fractional f) => (s -> f) -> Descriptor s f
+f32 f = Descriptor (realToFrac <$> getFloat32le, \s' -> putFloat32le (realToFrac $ f s') >> pure (f s'))
 
-i64 :: (a -> Int64) -> Descriptor a Int64
-i64 f = Descriptor (getInt64le, \s' -> putInt64le (f s') >> pure (f s'))
+f64 :: (Real f, Fractional f) => (s -> f) -> Descriptor s f
+f64 f = Descriptor (realToFrac <$> getFloat64le, \s' -> putFloat64le (realToFrac $ f s') >> pure (f s'))
 
-f32 :: (a -> Float) -> Descriptor a Float
-f32 f = Descriptor (getFloat32le, \s' -> putFloat32le (f s') >> pure (f s'))
+newtype LE a = LE { unwrapLE :: a }
+             deriving (Show, Read, Num, Eq, Ord, Enum, Integral, Real, Fractional)
 
-f64 :: (a -> Double) -> Descriptor a Double
-f64 f = Descriptor (getFloat64le, \s' -> putFloat64le (f s') >> pure (f s'))
+instance Describe (LE Word16) where
+    describe = w16
+
+instance Describe (LE Word32) where
+    describe = w32
+
+instance Describe (LE Word64) where
+    describe = w64
+
+instance Describe (LE Int16) where
+    describe = i16
+
+instance Describe (LE Int32) where
+    describe = i32
+
+instance Describe (LE Int64) where
+    describe = i64
+
+instance Describe (LE Float) where
+    describe = f32
+
+instance Describe (LE Double) where
+    describe = f64