+++ src/Data/Serialize/Describe/Endianness.hs
@@ -1,9 +1,0 @@
-module Data.Serialize.Describe.Endianness(
-  LE(..),
-  BE(..)
-) where
-
-newtype LE a = LE { unwrapLE :: a }
-
-newtype BE a = BE { unwrapBE :: a }
-